// src/app/interface/global.interface.ts
export type TRole = "admin" | "user" | "worker";
export type TGender = "Male" | "Female" | "Others";
export type TAccountStatus = "active" | "deactivate" | "pending" | "rejected";
export type TWithdrawStatus = "approved" | "pending" | "rejected";
export type TDepositActiveStatus =
  | "active"
  | "deactivate"
  | "pending"
  | "rejected";